Upgrade tools and install parameter (#76)
* put upgrade docker into utils

* put checks into utils

* upgrade reload cmd

* update logic in install.sh

* fix path error

* path error

* update config in install shell

* typo

* use absolute path

* error fix

* add c-gen

+++ install.sh
@@ -10,6 +10,7 @@ help()
 cat << EOF
 Usage:
     help                            show help information
+    --update                        update crust node
     --registry {cn|en}              use registry to accelerate docker pull
 EOF
 exit 0
@@ -96,17 +97,37 @@ download_docker_images()
 install_crust_node()
 {
     log_info "--------------Install crust node-------------"
+    local bin_file=/usr/bin/crust
+     
+    if [ -f "$bin_file" ] && [ x"$update" == x"true" ]; then
+        echo "Update crust node"
+        rm $bin_file
+        rm -rf $installdir/scripts
+        cp -r $basedir/scripts $installdir/
+        local upgrade_pid=$(ps -ef | grep "/opt/crust/crust-node/scripts/upgrade.sh" | grep -v grep | awk '{print $2}')
+        if [ x"$upgrade_pid" != x"" ]; then
+            kill -9 $upgrade_pid
+            nohup $installdir/scripts/upgrade.sh &>$installdir/logs/upgrade.log &
+            if [ $? -ne 0 ]; then
+                log_err "[ERROR] Start crust-sworker upgrade failed"
+                return 1
+            fi
+        fi
+    else
+        if [ -f "$installdir/scripts/uninstall.sh" ]; then
+            echo "Uninstall old crust node"
+            $installdir/scripts/uninstall.sh
+        fi
 
-    echo "uninstall old crust node"
-    $scriptdir/uninstall.sh
-
-    echo "Install crust node data"
-    mkdir -p $installdir
-    cp -r $basedir/scripts $installdir/
-    cp -r $basedir/etc $installdir/
-    cp $basedir/config.yaml $installdir/
-
-    echo "Change some configurations"
+        echo "Install new crust node"
+        mkdir -p $installdir
+        mkdir -p $installdir/logs
+        cp -r $basedir/etc $installdir/
+        cp $basedir/config.yaml $installdir/
+        cp -r $basedir/scripts $installdir/
+    fi
+    
+    echo "Change crust node configurations"
     sed -i 's/en/'$region'/g' $installdir/etc/region.conf
 
     echo "Install crust command line tool"
@@ -121,20 +142,31 @@ if [ $(id -u) -ne 0 ]; then
     exit 1
 fi
 
-case "$1" in
-    --registry)
-        if [ x"$2" == x"" ] || [[ x"$2" != x"cn" && x"$2" != x"en" ]]; then
-            help
-        fi
+region="en"
+update="false"
 
-        region=$2
-        ;;
-    "")
-        region="en"
-        ;;
-    *)
-        help
-esac
+while true ; do
+    case "$1" in
+        --registry)
+            if [ x"$2" == x"" ] || [[ x"$2" != x"cn" && x"$2" != x"en" ]]; then
+                help
+            fi
+            region=$2
+            shift 2
+            ;;
+        --update)
+            update="true"
+            shift 1
+            ;;
+        "")
+            shift ;
+            break ;;
+        *)
+            help
+            break;
+            ;;
+    esac
+done
 
 install_depenencies
 download_docker_images
